Moved card swiper to home page
